pyproject.toml
[project]
name = "trendwatcher"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "rich>=13.0",
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "lxml>=5.0",
  "pyyaml>=6.0",
  "python-dateutil>=2.9",

  # Feature 1 & 3: LLM Analysis and Product Matching
  "anthropic>=0.40.0",

  # Feature 2: New Data Sources
  "praw>=7.7.1",
  "feedparser>=6.0.11",

  # Feature 4: Scheduling
  "apscheduler>=3.10.4",

  # Feature 5: Reporting
  "slack-sdk>=3.27.0",

  # Environment configuration
  "python-dotenv>=1.0.0",

  # Explicit dependencies (already used)
  "pytrends>=4.9.2",
  "rapidfuzz>=3.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
